Better code structure
- Break reconcilers into pod and cluster
